usb: Rename usb_cfg_data to usb_class_data.
The name 'usb_cfg_data' is legacy of initial implementation
of the USB device stack. This structure was intended to be used to
configure the device in run time by calling usb_set_config().

Because of how things grown this was used to represent the data
of classes rather than the device as a whole. One device may consist of
many classes (composite device). When the stack was initially
developed this was not considered.

Alongside with renaming usb_cfg_data -> usb_class_data, all relevant
functions/macros/variables are updated accordingly.

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>